Removes unnecessary files
