--- client/src/components/nav-bar.js
@@ -1,18 +1,0 @@
-import React from 'react';
-
-import AuthNav from './auth-nav';
-
-const NavBar = () => {
-  return (
-    <div className="nav-container mb-3">
-      <nav className="navbar navbar-expand-md navbar-light bg-light">
-        <div className="container">
-          <div className="navbar-brand logo" />
-          <AuthNav />
-        </div>
-      </nav>
-    </div>
-  );
-};
-
-export default NavBar;
